third_party/WebKit/LayoutTests/virtual/service-worker-navigation-preload-disabled/http/tests/serviceworker/chromium/navigation-preload-origin-trial-methods-expected.txt

CONSOLE WARNING: line 13: This test relies on console message comparison so there can be different -expected.txt files for virtual test suites.
CONSOLE MESSAGE: line 68: --normal SW--
disable() rejected with: AbortError: Failed to enable or disable navigation preload: The service worker script does not have a valid Navigation Preload Origin Trial token.
enable() rejected with: AbortError: Failed to enable or disable navigation preload: The service worker script does not have a valid Navigation Preload Origin Trial token.
getState() resolved with: {"enabled":false,"headerValue":"true"}
setHeaderValue() rejected with: AbortError: Failed to set navigation preload header: The service worker script does not have a valid Navigation Preload Origin Trial token.
log from SW
Activate event handler
registration.navigationPreload not found
Fetch event handler
event.preloadResponse not found

CONSOLE MESSAGE: line 79: --Origin-Trial enabled SW--
disable() rejected with: AbortError: Failed to enable or disable navigation preload: Navigation Preload is disabled by command line flag.
enable() rejected with: AbortError: Failed to enable or disable navigation preload: Navigation Preload is disabled by command line flag.
getState() resolved with: {"enabled":false,"headerValue":"true"}
setHeaderValue() rejected with: AbortError: Failed to set navigation preload header: Navigation Preload is disabled by command line flag.
log from SW
Activate event handler
disable() rejected with: AbortError: Failed to enable or disable navigation preload: Navigation Preload is disabled by command line flag.
enable() rejected with: AbortError: Failed to enable or disable navigation preload: Navigation Preload is disabled by command line flag.
getState() resolved with: {"enabled":false,"headerValue":"true"}
setHeaderValue() rejected with: AbortError: Failed to set navigation preload header: Navigation Preload is disabled by command line flag.
Fetch event handler
event.preloadResponse resolved with: undefined

This is a testharness.js-based test.
PASS Calling Navigation preload related methods for normal SW. 
PASS Calling Navigation preload related methods for Origin-Trial enabled SW. 
Harness: the test ran to completion.
